batch: Extract ownership update orchestration
Batch source refresh currently owns stale-source repair and the
ownership-update API used by command paths that add selected lines to a
batch.

That makes a source-refresh module the public entry point for translation
and merge orchestration, even when callers need ownership update
preparation rather than source repair.

This commit moves the prepared-update dataclass, ownership translation, and
acquisition context into batch.ownership_update, then retargets command
callers to that narrower API.

Follow-up test commits will retarget coverage and pin the boundary so
source refresh stays focused on stale-source repair.
