+++ src/git_stage_batch/commands/selection/selected_change_batch_staging.py
@@ -6,8 +6,8 @@ import sys
 
 from ...batch.source_annotation import annotate_with_batch_source
 from ...batch.lifecycle import create_batch
+from ...batch.ownership_update import acquire_batch_ownership_update_for_selection
 from ...batch.query import read_batch_metadata
-from ...batch.source_refresh import acquire_batch_ownership_update_for_selection
 from ...batch.text_file_storage import add_file_to_batch
 from ...batch.validation import batch_exists
 from ...core.diff_parser import (
